requirements.txt
Flask==3.0.3
waitress==3.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1